handout_from_DS_agent/requirements.txt
numpy==1.24.4
pandas==2.0.3
scikit-learn==1.5.1
xgboost==3.0.5
joblib==1.5.1
pyyaml==6.0.2
